specify core exports, move core/api to core/src, improve pat-info test setup
