+++ packages/core/src/api/gist.js
@@ -1,39 +1,32 @@
 // @ts-check
 
-import { renderWakatimeCard } from "../src/cards/wakatime.js";
+import { renderGistCard } from "../cards/gist.js";
 import {
   MissingParamError,
   retrieveSecondaryMessage,
-} from "../src/common/error.js";
-import { parseArray, parseBoolean } from "../src/common/ops.js";
-import { renderError } from "../src/common/render.js";
-import { fetchWakatimeStats } from "../src/fetchers/wakatime.js";
-import { isLocaleAvailable } from "../src/translations.js";
+} from "../common/error.js";
+import { parseBoolean } from "../common/ops.js";
+import { renderError } from "../common/render.js";
+import { fetchGist } from "../fetchers/gist.js";
+import { isLocaleAvailable } from "../translations.js";
 
 // @ts-ignore
-export default async ({
-  username,
-  title_color,
-  icon_color,
-  hide_border,
-  card_width,
-  line_height,
-  text_color,
-  bg_color,
-  theme,
-  hide_title,
-  hide_progress,
-  custom_title,
-  locale,
-  layout,
-  langs_count,
-  hide,
-  api_domain,
-  border_radius,
-  border_color,
-  display_format,
-  disable_animations,
-}) => {
+export default async (
+  {
+    id,
+    title_color,
+    icon_color,
+    text_color,
+    bg_color,
+    theme,
+    locale,
+    border_radius,
+    border_color,
+    show_owner,
+    hide_border,
+  },
+  pat = null,
+) => {
   if (locale && !isLocaleAvailable(locale)) {
     return {
       status: "error - permanent",
@@ -52,30 +45,21 @@ export default async ({
   }
 
   try {
-    const stats = await fetchWakatimeStats({ username, api_domain });
+    const gistData = await fetchGist(id, pat);
 
     return {
       status: "success",
-      content: renderWakatimeCard(stats, {
-        custom_title,
-        hide_title: parseBoolean(hide_title),
-        hide_border: parseBoolean(hide_border),
-        card_width: parseInt(card_width, 10),
-        hide: parseArray(hide),
-        line_height,
+      content: renderGistCard(gistData, {
         title_color,
         icon_color,
         text_color,
         bg_color,
         theme,
-        hide_progress,
         border_radius,
         border_color,
         locale: locale ? locale.toLowerCase() : null,
-        layout,
-        langs_count,
-        display_format,
-        disable_animations: parseBoolean(disable_animations),
+        show_owner: parseBoolean(show_owner),
+        hide_border: parseBoolean(hide_border),
       }),
     };
   } catch (err) {
